+++ sdk/src/androidTest/java/com/punchthrough/bean/sdk/TestBeanSimple.java
@@ -1,9 +1,9 @@
 package com.punchthrough.bean.sdk;
 
-import com.punchthrough.bean.sdk.TestingUtils.LooperRunner;
+import com.punchthrough.bean.sdk.util.TestingUtils.LooperRunner;
+import com.punchthrough.bean.sdk.util.TestingUtils.BeanUtils;
 
 import android.test.AndroidTestCase;
-import android.test.suitebuilder.annotation.Suppress;
 
 import com.punchthrough.bean.sdk.message.BeanError;
 import com.punchthrough.bean.sdk.message.Callback;
@@ -11,104 +11,36 @@ import com.punchthrough.bean.sdk.message.DeviceInfo;
 import com.punchthrough.bean.sdk.message.ScratchBank;
 import com.punchthrough.bean.sdk.message.ScratchData;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Integration tests for the Bean.
- * <p/>
+ *
  * Prerequisites:
- * - Bean within range
+ * - Bean within range named TESTBEAN
  * - Android device connected over USB
  */
-public class TestBean extends AndroidTestCase {
+public class TestBeanSimple extends AndroidTestCase {
 
-    private LooperRunner lr = new LooperRunner();
+    private LooperRunner lr = new LooperRunner(BeanManager.getInstance().mHandler.getLooper());
     private Thread lrThread = new Thread(lr);
 
     protected void setUp() {
         lrThread.start();
     }
 
-    protected void tearDown() throws InterruptedException {
-        /*
-        * "Speaking very generally, never quit() your looper threads. That method exists
-        * mostly for historical and testing reasons. In Real Life™, I recommend that you
-        * continue to reuse the same looper thread(s) for the life of the process rather
-        * than creating/quitting them."
-        * */
-//        lr.quit();
-//        lrThread.join();
-    }
+    protected void tearDown() throws InterruptedException {}
 
-    private List<Bean> getBeans(int num) throws InterruptedException {
-        final CountDownLatch beanLatch = new CountDownLatch(num);
-        final List<Bean> beans = new ArrayList<>();
-
-        BeanDiscoveryListener listener = new BeanDiscoveryListener() {
-            @Override
-            public void onBeanDiscovered(Bean bean, int rssi) {
-                beans.add(bean);
-                beanLatch.countDown();
-            }
-
-            @Override
-            public void onDiscoveryComplete() {
-                beanLatch.countDown();
-            }
-        };
-
-        boolean startedOK = BeanManager.getInstance().startDiscovery(listener);
-        assertThat(startedOK).isTrue();
-        beanLatch.await(60, TimeUnit.SECONDS);
-        assertThat(beans.size()).isEqualTo(num);
-        return beans;
-    }
-
-    private Bean getBeanByName(String name) throws InterruptedException {
-        final CountDownLatch beanLatch = new CountDownLatch(1);
-        final List<Bean> beans = new ArrayList<>();
-
-        final String targetName = name;
-
-        BeanDiscoveryListener listener = new BeanDiscoveryListener() {
-            @Override
-            public void onBeanDiscovered(Bean bean, int rssi) {
-                if (bean.getDevice().getName().equals(targetName)) {
-                    System.out.println("Found Bean!!!!!!!");
-                    beans.add(bean);
-                    beanLatch.countDown();
-                }
-            }
-
-            @Override
-            public void onDiscoveryComplete() {
-                System.out.println("Nothing");
-                beanLatch.countDown();
-            }
-        };
-
-        boolean startedOK = BeanManager.getInstance().startDiscovery(listener);
-        assertThat(startedOK).isTrue();
-        beanLatch.await(60, TimeUnit.SECONDS);
-        if (beans.isEmpty()) {
-            fail("Couldn't find bean by name: " + name);
-        }
-        return beans.get(0);
-    }
-
-    public void testBeanDeviceInfo() throws InterruptedException {
+    public void testBeanDeviceInfo() throws Exception {
         /** Read device information from a bean
          *
          * Warning: This test requires a nearby bean named "TESTBEAN"
          */
 
-        Bean bean = this.getBeanByName("TESTBEAN");
+        Bean bean = BeanUtils.getBeanByName("TESTBEAN");
         final CountDownLatch connectionLatch = new CountDownLatch(1);
         final HashMap testState = new HashMap();
         testState.put("connected", false);
@@ -164,12 +96,12 @@ public class TestBean extends AndroidTestCase {
         }
     }
 
-    public void testBeanReadWriteScratchBank() throws InterruptedException {
+    public void testBeanReadWriteScratchBank() throws Exception {
         /** Test Scratch characteristic functionality
          *
          * Warning: This test requires a nearby bean named "TESTBEAN"
          */
-        Bean bean = this.getBeanByName("TESTBEAN");
+        Bean bean = BeanUtils.getBeanByName("TESTBEAN");
         final CountDownLatch connectionLatch = new CountDownLatch(1);
         final HashMap testState = new HashMap();
         testState.put("connected", false);
@@ -189,6 +121,7 @@ public class TestBean extends AndroidTestCase {
 
             @Override
             public void onDisconnected() {
+                System.out.println("Disconnected");
             }
 
             @Override
@@ -257,68 +190,4 @@ public class TestBean extends AndroidTestCase {
         }
     }
 
-    @Suppress
-    public void testConnectMultipleBeansWithSameListener() throws InterruptedException {
-        /* This test requires at least 3 beans nearby to pass */
-
-        final List<Bean> beans = this.getBeans(3);
-        final Bean beanA = beans.get(0);
-        final Bean beanB = beans.get(1);
-        final Bean beanC = beans.get(2);
-        final CountDownLatch connectionLatch = new CountDownLatch(2);
-        final HashMap<String, Boolean> connectionState = new HashMap<>();
-        connectionState.put("bean_a_connected", false);
-        connectionState.put("bean_b_connected", false);
-
-        BeanListener beanListener = new BeanListener() {
-
-            @Override
-            public void onConnected() {
-                if (connectionState.get("bean_a_connected") == false) {
-                    if (beanA.isConnected()) {
-                        connectionState.put("bean_a_connected", true);
-                        connectionLatch.countDown();
-                    }
-                }
-
-                if (connectionState.get("bean_b_connected") == false) {
-                    if (beanB.isConnected()) {
-                        connectionState.put("bean_b_connected", true);
-                        connectionLatch.countDown();
-                    }
-                }
-
-                if (beanC.isConnected()) {
-                    fail("Bean C not suppose to connect!");
-                }
-            }
-
-            @Override
-            public void onConnectionFailed() {
-                fail("Connection failed!");
-            }
-
-            @Override
-            public void onDisconnected() {
-            }
-
-            @Override
-            public void onSerialMessageReceived(byte[] data) {
-            }
-
-            @Override
-            public void onScratchValueChanged(ScratchBank bank, byte[] value) {
-            }
-
-            @Override
-            public void onError(BeanError error) {
-                fail(error.toString());
-            }
-        };
-
-        beanA.connect(getContext(), beanListener);
-        beanB.connect(getContext(), beanListener);
-        connectionLatch.await(60, TimeUnit.SECONDS);
-        // No need to assert anything, implicit success based on connection latch
-    }
 }