+++ sdk/src/main/java/com/punchthrough/bean/sdk/internal/serial/GattSerialTransportProfile.java
@@ -58,12 +58,18 @@ public class GattSerialTransportProfile extends BaseProfile {
         }
     };
     private List<GattSerialPacket> mPendingPackets = new ArrayList<>(32);
-    private Handler mHandler = new Handler(Looper.getMainLooper());
+    private Handler mHandler;
     private int mOutgoingMessageCount = 0;
     private MessageAssembler mMessageAssembler = new MessageAssembler();
 
     public GattSerialTransportProfile(GattClient client) {
         super(client);
+        mHandler = new Handler(Looper.getMainLooper());
+    }
+
+    public GattSerialTransportProfile(GattClient client, Handler handler) {
+        super(client);
+        mHandler = handler;
     }
 
     @Override
@@ -155,6 +161,7 @@ public class GattSerialTransportProfile extends BaseProfile {
             // scratch
             int index = BEAN_SCRATCH_UUIDS.indexOf(characteristic.getUuid());
             if (index > -1) {
+                index += 1;
                 if (BuildConfig.DEBUG) {
                     Log.d(TAG, "Received scratch bank update (" + index + ")");
                 }